+++ packages/bot-utils/tests/help.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { generateHelp } from '../help.js';
+import { generateHelp } from '../src/help.js';
 
 describe('help', () => {
 	describe('generateHelp', () => {
