+++ packages/bot-utils/tests/response.test.ts
@@ -5,8 +5,8 @@ import {
 	LinkBlockType,
 	responseHeader,
 	shareMessageFooter,
-} from '../response.js';
-import { ProbeMeasurement } from '../types.js';
+} from '../src/response.js';
+import { ProbeMeasurement } from '../src/types.js';
 
 describe('Response', () => {
 	const boldSeparator = '*';
