+++ packages/discord/tests/utils.ts
@@ -1,6 +1,6 @@
 import { vi } from 'vitest';
 
-import probeData from '@globalping/bot-utils/src/tests/mocks/probedata.json' assert { type: 'json' };
+import probeData from '@globalping/bot-utils/dist/tests/mocks/probedata.json' assert { type: 'json' };
 import { Logger, Measurement } from '@globalping/bot-utils';
 import { ChatInputCommandInteraction } from 'discord.js';
