+++ packages/main/package.json
@@ -2,13 +2,13 @@
 	"name": "main",
 	"private": true,
 	"type": "module",
-	"main": "./dist/main.js",
-	"types": "./dist/main.d.ts",
+	"main": "./dist/src/main.js",
+	"types": "./dist/src/main.d.ts",
 	"scripts": {
 		"build": "tsc",
 		"start": "NODE_ENV=production node dist/main.js",
 		"discord:deploy-commands": "NODE_ENV=production tsx src/discord/deploy-commands.ts",
-		"dev": "tsx src/main.ts",
+		"start:dev": "tsx src/main.ts",
 		"watch": "tsx watch src/main.ts"
 	},
 	"dependencies": {
