+++ tsconfig.json
@@ -31,11 +31,12 @@
 		]
 	},
 	"include": [
-		"packages/**/*"
+		"packages/**/*",
+		"packages/**/*.json"
 	],
 	"exclude": [
-		"node_modules",
-		"dist"
+		"dist",
+		"node_modules"
 	],
 	"references": [
 		{
